client/requirements.txt
ansicon==1.89.0
blessed==1.17.6
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jinxed==1.0.1
prompt-toolkit==1.0.14
pygame==2.0.1
Pygments==2.7.4
PyInquirer==1.0.3
python-editor==1.0.4
readchar==2.0.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
urllib3==1.26.2
wcwidth==0.2.5
websocket-client==0.57.0